# src/openapi/openapi.json
{
    "openapi":"3.0.1","info":{"title":"OpenApi specification - TPS","description":"OpenApi documentation for Spring Security","termsOfService":"Terms of service","contact":{"name":"TPS","email":"test@tps.com"},"license":{"name":"Licence name","url":"https://some-url.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8089/api/v1","description":"Local ENV"},{"url":"http://localhost:8088/api/v1","description":"Local ENV_2"}],"security":[{"bearerAuth":[]}],"paths":{"/users/{userId}/roles/{roleId}":{"put":{"tags":["User Management"],"operationId":"changeUserRole","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"roleId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/users/{id}":{"get":{"tags":["User Management"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["User Management"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"delete":{"tags":["User Management"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"204":{"description":"No Content"}}}},"/roles/{id}":{"get":{"tags":["Role Management"],"operationId":"getRoleById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Role"}}}}}},"put":{"tags":["Role Management"],"operationId":"updateRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Role"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Role"}}}}}}},"/projects/{id}":{"put":{"tags":["project-controller"],"operationId":"updateProject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}},"delete":{"tags":["project-controller"],"operationId":"deleteProject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/projects/{id_project}/{id_user}":{"put":{"tags":["project-controller"],"operationId":"addUserToProject","parameters":[{"name":"id_project","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id_user","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/projects/status/{id}":{"put":{"tags":["project-controller"],"operationId":"updateProjectStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/products/{id}":{"put":{"tags":["product-controller"],"operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}},"delete":{"tags":["product-controller"],"operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/presence/{presenceId}":{"put":{"tags":["presence-controller"],"operationId":"updatePresence","parameters":[{"name":"presenceId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PresenceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/orders/{id}":{"put":{"tags":["order-controller"],"operationId":"updateOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}},"delete":{"tags":["order-controller"],"operationId":"deleteOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/users":{"get":{"tags":["User Management"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["User Management"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"201":{"description":"Created"}}}},"/roles":{"get":{"tags":["Role Management"],"operationId":"getAllRoles","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Role"}}}}}}},"post":{"tags":["Role Management"],"operationId":"createRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Role"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Role"}}}}}}},"/projects":{"get":{"tags":["project-controller"],"operationId":"findAllProjects","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectResponse"}}}}}}},"post":{"tags":["project-controller"],"operationId":"saveProject","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/products":{"get":{"tags":["product-controller"],"operationId":"findAllProducts","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"post":{"tags":["product-controller"],"operationId":"saveProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/presence/project/{projectId}":{"get":{"tags":["presence-controller"],"operationId":"findAllPresences","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PresenceResponse"}}}}}}},"post":{"tags":["presence-controller"],"operationId":"savePresence","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PresenceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/orders":{"get":{"tags":["order-controller"],"operationId":"findAllOrders","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderResponse"}}}}}}},"post":{"tags":["order-controller"],"operationId":"saveOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/demandes/reject/{id}":{"post":{"tags":["demande-controller"],"operationId":"rejectDemande","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Demande"}}}}}}},"/demandes/create":{"post":{"tags":["demande-controller"],"operationId":"createDemande","parameters":[{"name":"projectId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"productIds","in":"query","required":true,"schema":{"type":"array","items":{"type":"integer","format":"int32"}}},{"name":"quantity","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Demande"}}}}}}},"/demandes/approve/{id}":{"post":{"tags":["demande-controller"],"operationId":"approveDemande","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Demande"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/projects/{project-id}":{"get":{"tags":["project-controller"],"operationId":"findProjectById","parameters":[{"name":"project-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}}}}},"/products/{product-id}":{"get":{"tags":["product-controller"],"operationId":"findProductById","parameters":[{"name":"product-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/presence/by-date-and-project":{"get":{"tags":["presence-controller"],"operationId":"getPresenceByDateAndProject","parameters":[{"name":"date","in":"query","required":true,"schema":{"type":"string"}},{"name":"projectId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Presence"}}}}}}}},"/orders/{order-id}":{"get":{"tags":["order-controller"],"operationId":"findOrderById","parameters":[{"name":"order-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}}}},"/auth/activate-account":{"get":{"tags":["Authentication"],"operationId":"confirm","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Demande":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"},"entry_date":{"type":"string","format":"date"},"sort_date":{"type":"string","format":"date"},"status":{"type":"string"},"project":{"$ref":"#/components/schemas/Project"},"products":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Product":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"type":{"type":"string"},"unit_price":{"type":"integer","format":"int32"},"unity":{"type":"string"},"demandes":{"type":"array","items":{"$ref":"#/components/schemas/Demande"}}}},"Project":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"createdBy":{"type":"string"},"description":{"type":"string"},"type":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"end_date":{"type":"string","format":"date-time"},"budget":{"type":"integer","format":"int32"},"status":{"type":"string"},"location":{"type":"string"},"users":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"demandes":{"type":"array","items":{"$ref":"#/components/schemas/Demande"}}}},"Role":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"firstname":{"type":"string"},"lastname":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"email":{"type":"string"},"password":{"type":"string"},"accountLocked":{"type":"boolean"},"enabled":{"type":"boolean"},"phone":{"type":"string"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/Role"}},"projects":{"type":"array","items":{"$ref":"#/components/schemas/Project"}},"createdDate":{"type":"string","format":"date"},"name":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date"},"username":{"type":"string"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"accountNonExpired":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"}}},"ProjectRequest":{"required":["budget","created_At","location","name","status","type"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"endDate":{"type":"string","format":"date-time"},"description":{"type":"string"},"type":{"type":"string"},"created_At":{"type":"string","format":"date-time"},"budget":{"type":"integer","format":"int32"},"status":{"type":"string"},"location":{"type":"string"},"userIds":{"type":"array","items":{"type":"integer","format":"int32"}}}},"ProductRequest":{"required":["name","type","unit_price","unity"],"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string"},"unit_price":{"type":"integer","format":"int32"},"unity":{"type":"string"}}},"PresenceRequest":{"required":["company","personalName","type"],"type":"object","properties":{"personalName":{"type":"string"},"type":{"type":"string"},"company":{"type":"string"},"subtype":{"type":"string"},"presenceDescription":{"type":"string"},"costs":{"type":"integer","format":"int32"}}},"OrderItemRequest":{"type":"object","properties":{"productId":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"}}},"OrderRequest":{"type":"object","properties":{"orderDate":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int32"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemRequest"}}}},"RegistrationRequest":{"required":["email","firstname","lastname","password"],"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}},"ProjectResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"type":{"type":"string"},"created_At":{"type":"string","format":"date-time"},"budget":{"type":"integer","format":"int32"},"status":{"type":"string"},"location":{"type":"string"},"users":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"createdBy":{"type":"string"}}},"ProductResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"type":{"type":"string"},"unit_price":{"type":"integer","format":"int32"},"unity":{"type":"string"}}},"PresenceResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"personalName":{"type":"string"},"company":{"type":"string"},"type":{"type":"string"},"subtype":{"type":"string"},"presenceDescription":{"type":"string"},"costs":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date"}}},"Presence":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"personalName":{"type":"string"},"company":{"type":"string"},"type":{"type":"string"},"subtype":{"type":"string"},"presenceDescription":{"type":"string"},"costs":{"type":"integer","format":"int32"},"createdBy":{"type":"string"},"createdAt":{"type":"string","format":"date"}}},"OrderItemResponse":{"type":"object","properties":{"productId":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"}}},"OrderResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"orderDate":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int32"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemResponse"}}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}
}